--- src/components/UI/molecules/molecules-main/LoginBtn.jsx
@@ -1,11 +1,0 @@
-import Button from '../../atoms/atoms-main/Button';
-import TextStyle from '../../atoms/atoms-main/TextStyle';
-
-const LoginBtn = ({ name }) => {
-	return (
-		<Button submitBtn>
-			<TextStyle whiteNormalBold>{name}</TextStyle>{' '}
-		</Button>
-	);
-};
-export default LoginBtn;
