+++ src/components/UI/molecules/molecules-main/RoomsLink.jsx
@@ -4,7 +4,7 @@ import Modal from '../../../../portal/Modal';
 // import Modal from '../../../../potals/modals/Modal';
 import Button from '../../atoms/atoms-main/Button';
 import TextStyle from '../../atoms/atoms-main/TextStyle';
-import LoginModalContainer from '../../../../portal/modals/LoginModalContainer';
+import AuthModalContainer from '../../../../containers/AuthModalContainer';
 
 const RoomsLink = (props) => {
   console.log('dd', props);
@@ -26,7 +26,7 @@ const RoomsLink = (props) => {
       </div>
       {modal && (
         <Modal>
-          <LoginModalContainer modal={modal} setModal={setModal} />
+          <AuthModalContainer modal={modal} setModal={setModal} />
         </Modal>
       )}
     </>
