+++ src/components/UI/molecules/molecules-modals-auth/EmailLoginSubmit.jsx
@@ -1,0 +1,34 @@
+import React from 'react';
+import styled from 'styled-components';
+import Button from '../../atoms/atoms-main/Button';
+import CircleDiv from '../../atoms/atoms-main/DivStyle';
+import TextStyle from '../../atoms/atoms-main/TextStyle';
+import AuthBtn from './AuthBtn';
+
+const EmailLoginStyle = styled.div`
+  width: 400px;
+
+  display: flex;
+  flex-direction: column;
+  .search-password {
+    border: 0;
+    margin-bottom: 2%;
+  }
+`;
+
+const EmailLoginSubmit = () => {
+  return (
+    <EmailLoginStyle>
+      <CircleDiv className="submit">
+        <CircleDiv className="submit-div">
+          <Button className="search-password">
+            <TextStyle greentextLine>비밀번호를 잊으셨나요?</TextStyle>
+          </Button>
+
+          <AuthBtn name="로그인" />
+        </CircleDiv>
+      </CircleDiv>
+    </EmailLoginStyle>
+  );
+};
+export default EmailLoginSubmit;
