+++ src/portal/Modal.js
@@ -1,5 +1,6 @@
 import ReactDOM from 'react-dom';
 
-const Modal = ({ children }) => ReactDOM.createPortal(children, document.getElementById('modal'));
+const Modal = ({ children }) =>
+  ReactDOM.createPortal(children, document.getElementById('modal'));
 
 export default Modal;
